nerdtalk/requirements.txt
astroid==1.5.3
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.5
django-social-share==1.1.2
django-tinymce==2.7.0
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.0.6
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.7.3
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.1
wrapt==1.10.11
